add all remappings and tests for echo

+++ test/echo/libraries/LibAppStorageUpgradeTest.sol
@@ -1,10 +1,10 @@
 // SPDX-License-Identifier: MIT
 pragma solidity 0.8.18;
 
-import "../../interfaces/IIngesterRegistration.sol";
-import "../../interfaces/IIngesterGroupManager.sol";
-import "../../interfaces/IIngesterDataGathering.sol";
-import {LibDiamond} from "../../libraries/LibDiamond.sol";
+import "../../../src/echo/contracts/interfaces/IIngesterRegistration.sol";
+import "../../../src/echo/contracts/interfaces/IIngesterGroupManager.sol";
+import "../../../src/echo/contracts/interfaces/IIngesterDataGathering.sol";
+import {LibDiamond} from "../../../src/echo/contracts/libraries/LibDiamond.sol";
 
 struct AppStorageTest {
     //Registry
@@ -17,7 +17,7 @@ struct AppStorageTest {
     uint256 maxClusterSize;
     uint256 maxGroupsPerIngester;
     uint256 maxIngestersPerGroup;
-    mapping(uint => IIngesterGroupManager.Cluster) ingesterClusters;
+    mapping(uint256 => IIngesterGroupManager.Cluster) ingesterClusters;
     uint256[] clusterIds;
     mapping(string => IIngesterGroupManager.Group) groups;
     string[] groupUsernames;
@@ -34,18 +34,9 @@ library LibAppStorageTest {
     bytes32 internal constant CONTROLLER_ROLE = keccak256("CONTROLLER_ROLE");
 
     event UnAllocatedGroupsAdded(string[] groups);
-    event IngesterRemovedFromGroup(
-        address indexed ingesterAddress,
-        string indexed groupId
-    );
-    event IngesterRegisteredGroups(
-        address indexed ingesterAddress,
-        string[] assignedGroups
-    );
-    event IngesterAddedToCluster(
-        address indexed ingesterAddress,
-        uint256 indexed clusterId
-    );
+    event IngesterRemovedFromGroup(address indexed ingesterAddress, string indexed groupId);
+    event IngesterRegisteredGroups(address indexed ingesterAddress, string[] assignedGroups);
+    event IngesterAddedToCluster(address indexed ingesterAddress, uint256 indexed clusterId);
 
     function appStorage() internal pure returns (AppStorageTest storage ds) {
         assembly {
@@ -87,10 +78,7 @@ library LibAppStorageTest {
         return ingesterAddresses;
     }
 
-    function addIngesterToCluster(
-        address ingesterAddress,
-        address controllerAddress
-    ) internal returns (uint256) {
+    function addIngesterToCluster(address ingesterAddress, address controllerAddress) internal returns (uint256) {
         AppStorageTest storage s = appStorage();
 
         uint256 clusterId = 0;
@@ -102,24 +90,19 @@ library LibAppStorageTest {
 
         s.ingesterClusters[clusterId].ingesterAddresses.push(ingesterAddress);
         require(
-            (s.maxGroupsPerIngester *
-                s.ingesterClusters[clusterId].ingesterAddresses.length) >=
-                s.ingesterClusters[clusterId].clusterGroupCount,
+            (s.maxGroupsPerIngester * s.ingesterClusters[clusterId].ingesterAddresses.length)
+                >= s.ingesterClusters[clusterId].clusterGroupCount,
             "More groups in cluster than cluster constraints"
         );
-        s.ingesterClusters[clusterId].clusterRemainingCapacity =
-            (s.maxGroupsPerIngester *
-                s.ingesterClusters[clusterId].ingesterAddresses.length) -
-            s.ingesterClusters[clusterId].clusterGroupCount;
+        s.ingesterClusters[clusterId].clusterRemainingCapacity = (
+            s.maxGroupsPerIngester * s.ingesterClusters[clusterId].ingesterAddresses.length
+        ) - s.ingesterClusters[clusterId].clusterGroupCount;
 
         emit IngesterAddedToCluster(ingesterAddress, clusterId);
         return clusterId;
     }
 
-    function getAvailableCluster(
-        address ingesterAddress,
-        address controllerAddress
-    ) internal returns (uint256) {
+    function getAvailableCluster(address ingesterAddress, address controllerAddress) internal returns (uint256) {
         AppStorageTest storage s = appStorage();
 
         uint256 availableCluster = 0;
@@ -134,42 +117,24 @@ library LibAppStorageTest {
 
         //otherwise go through the existing clusters and get the first available one
         for (uint256 i = 0; i < numClusters; i++) {
-            if (
-                s.ingesterClusters[s.clusterIds[i]].ingesterAddresses.length <
-                s.maxClusterSize
-            ) {
+            if (s.ingesterClusters[s.clusterIds[i]].ingesterAddresses.length < s.maxClusterSize) {
                 availableCluster = s.clusterIds[i];
                 foundAvailableCluster = true;
 
                 //If there is Duplication.
                 //Restrict more than one controller wallet owning one ingester per cluster
                 if (s.maxIngestersPerGroup > 1) {
-                    IIngesterRegistration.Ingester[]
-                        memory controllerIngesters = s.controllerToIngesters[
-                            controllerAddress
-                        ];
+                    IIngesterRegistration.Ingester[] memory controllerIngesters =
+                        s.controllerToIngesters[controllerAddress];
 
-                    for (
-                        uint j = 0;
-                        j <
-                        s
-                            .ingesterClusters[availableCluster]
-                            .ingesterAddresses
-                            .length;
-                        j++
-                    ) {
-                        for (uint k = 0; k < controllerIngesters.length; k++) {
+                    for (uint256 j = 0; j < s.ingesterClusters[availableCluster].ingesterAddresses.length; j++) {
+                        for (uint256 k = 0; k < controllerIngesters.length; k++) {
                             //controller ingester addresses match ingesterAddresses within the cluster and it's not the current ingester we are trying to assign
                             //continue looking for another cluster
                             if (
-                                s
-                                    .ingesterClusters[availableCluster]
-                                    .ingesterAddresses[j] ==
-                                controllerIngesters[k].ingesterAddress &&
-                                s
-                                    .ingesterClusters[availableCluster]
-                                    .ingesterAddresses[j] !=
-                                ingesterAddress
+                                s.ingesterClusters[availableCluster].ingesterAddresses[j]
+                                    == controllerIngesters[k].ingesterAddress
+                                    && s.ingesterClusters[availableCluster].ingesterAddresses[j] != ingesterAddress
                             ) {
                                 foundAvailableCluster = false;
                             }
@@ -193,34 +158,18 @@ library LibAppStorageTest {
         }
     }
 
-    function removeIngesterFromGroups(
-        uint256 clusterId,
-        address ingesterAddress
-    ) internal {
+    function removeIngesterFromGroups(uint256 clusterId, address ingesterAddress) internal {
         AppStorageTest storage s = LibAppStorageTest.appStorage();
 
-        string[] memory groups = s
-            .ingesterClusters[clusterId]
-            .ingesterToAssignedGroups[ingesterAddress];
+        string[] memory groups = s.ingesterClusters[clusterId].ingesterToAssignedGroups[ingesterAddress];
         for (uint256 i = 0; i < groups.length; ++i) {
             // These arrays are capped by the s.maxNumberIngesterPerGroup
-            uint256 amountOfIngestersPerGroup = s
-                .groups[groups[i]]
-                .ingesterAddresses
-                .length;
+            uint256 amountOfIngestersPerGroup = s.groups[groups[i]].ingesterAddresses.length;
             for (uint256 j = 0; j < amountOfIngestersPerGroup; j++) {
-                if (
-                    s.groups[groups[i]].ingesterAddresses[j] == ingesterAddress
-                ) {
+                if (s.groups[groups[i]].ingesterAddresses[j] == ingesterAddress) {
                     //delete ingesterAddress from groups and readjust array length so s.maxNumberIngesterPerGroup check remains truthful
-                    for (
-                        uint256 z = j;
-                        z < amountOfIngestersPerGroup - 1;
-                        ++z
-                    ) {
-                        s.groups[groups[i]].ingesterAddresses[z] = s
-                            .groups[groups[i]]
-                            .ingesterAddresses[z + 1];
+                    for (uint256 z = j; z < amountOfIngestersPerGroup - 1; ++z) {
+                        s.groups[groups[i]].ingesterAddresses[z] = s.groups[groups[i]].ingesterAddresses[z + 1];
                     }
                     s.groups[groups[i]].ingesterAddresses.pop();
                     emit IngesterRemovedFromGroup(ingesterAddress, groups[i]);
@@ -230,66 +179,39 @@ library LibAppStorageTest {
         }
     }
 
-    function removeIngesterFromCluster(
-        address ingesterAddress,
-        uint256 clusterId
-    ) internal {
+    function removeIngesterFromCluster(address ingesterAddress, uint256 clusterId) internal {
         AppStorageTest storage s = LibAppStorageTest.appStorage();
 
-        uint256 numIngesters = s
-            .ingesterClusters[clusterId]
-            .ingesterAddresses
-            .length;
+        uint256 numIngesters = s.ingesterClusters[clusterId].ingesterAddresses.length;
         uint256 ingesterIndexToRemove = 0;
         for (uint256 i = 0; i < numIngesters; i++) {
-            if (
-                s.ingesterClusters[clusterId].ingesterAddresses[i] ==
-                ingesterAddress
-            ) {
+            if (s.ingesterClusters[clusterId].ingesterAddresses[i] == ingesterAddress) {
                 ingesterIndexToRemove = i;
                 //Re-adjust the ingesterAddresses array
                 if (i != numIngesters - 1) {
-                    address ingesterAddressToMove = s
-                        .ingesterClusters[clusterId]
-                        .ingesterAddresses[numIngesters - 1];
-                    s.ingesterClusters[clusterId].ingesterAddresses[
-                        i
-                    ] = ingesterAddressToMove;
+                    address ingesterAddressToMove = s.ingesterClusters[clusterId].ingesterAddresses[numIngesters - 1];
+                    s.ingesterClusters[clusterId].ingesterAddresses[i] = ingesterAddressToMove;
                 }
                 s.ingesterClusters[clusterId].ingesterAddresses.pop();
 
-                if (
-                    s.ingesterClusters[clusterId].ingesterAddresses.length == 0
-                ) {
+                if (s.ingesterClusters[clusterId].ingesterAddresses.length == 0) {
                     removeCluster(clusterId);
                     delete s.ingesterClusters[clusterId];
                 } else {
                     //remove all assignedGroups from ingester within the cluster
-                    s.ingesterClusters[clusterId].clusterGroupCount -= s
-                        .ingesterClusters[clusterId]
-                        .ingesterToAssignedGroups[ingesterAddress]
-                        .length;
-                    s.ingesterClusters[clusterId].ingesterToAssignedGroups[
-                        ingesterAddress
-                    ] = new string[](0);
+                    s.ingesterClusters[clusterId].clusterGroupCount -=
+                        s.ingesterClusters[clusterId].ingesterToAssignedGroups[ingesterAddress].length;
+                    s.ingesterClusters[clusterId].ingesterToAssignedGroups[ingesterAddress] = new string[](0);
 
                     //re-calculate cluster remaining capacity
                     require(
-                        (s.maxGroupsPerIngester *
-                            s
-                                .ingesterClusters[clusterId]
-                                .ingesterAddresses
-                                .length) >=
-                            s.ingesterClusters[clusterId].clusterGroupCount,
+                        (s.maxGroupsPerIngester * s.ingesterClusters[clusterId].ingesterAddresses.length)
+                            >= s.ingesterClusters[clusterId].clusterGroupCount,
                         "More groups in cluster than cluster constraints"
                     );
-                    s.ingesterClusters[clusterId].clusterRemainingCapacity =
-                        (s.maxGroupsPerIngester *
-                            s
-                                .ingesterClusters[clusterId]
-                                .ingesterAddresses
-                                .length) -
-                        s.ingesterClusters[clusterId].clusterGroupCount;
+                    s.ingesterClusters[clusterId].clusterRemainingCapacity = (
+                        s.maxGroupsPerIngester * s.ingesterClusters[clusterId].ingesterAddresses.length
+                    ) - s.ingesterClusters[clusterId].clusterGroupCount;
                     break;
                 }
             }
